app/requirements.txt
# flask==2.2.5
# flask-socketio==5.3.3
# python-socketio==5.7.2
# eventlet==0.33.3
# flask-cors==3.0.10
# kafka-python

fastapi 
uvicorn[standard] 
pydantic
python-socketio 
kafka-python 
jinja2